+++ src/main/java/wooteco/subway/controller/SectionController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.dto.SectionSaveRequest;
+import wooteco.subway.dto.section.SectionSaveRequest;
 import wooteco.subway.service.SectionService;
 
 @RestController
