+++ src/main/java/wooteco/subway/dao/line/InmemoryLineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.line;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
