+++ src/main/java/wooteco/subway/dao/line/JdbcLineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.line;
 
 import java.sql.PreparedStatement;
 import java.util.List;
