+++ src/main/java/wooteco/subway/dao/station/InmemoryStationDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.station;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
