+++ src/main/java/wooteco/subway/dao/station/StationDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.station;
 
 import java.util.List;
 import wooteco.subway.domain.Station;
