+++ src/main/java/wooteco/subway/dto/line/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.line;
 
 import javax.validation.constraints.NotBlank;
 import wooteco.subway.domain.Line;
