+++ src/test/java/wooteco/subway/acceptance/SectionAcceptanceTest.java
@@ -11,11 +11,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.LineResponse;
-import wooteco.subway.dto.LineSaveRequest;
-import wooteco.subway.dto.SectionSaveRequest;
-import wooteco.subway.dto.StationResponse;
-import wooteco.subway.dto.StationSaveRequest;
+import wooteco.subway.dto.line.LineResponse;
+import wooteco.subway.dto.line.LineSaveRequest;
+import wooteco.subway.dto.section.SectionSaveRequest;
+import wooteco.subway.dto.station.StationResponse;
+import wooteco.subway.dto.station.StationSaveRequest;
 
 public class SectionAcceptanceTest extends AcceptanceTest {
