+++ src/test/java/wooteco/subway/dao/section/InmemorySectionDaoTest.java
@@ -1,10 +1,12 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.section;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.dao.line.InmemoryLineDao;
+import wooteco.subway.dao.station.InmemoryStationDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
