+++ src/test/java/wooteco/subway/dao/station/JdbcStationDaoTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.dao.station;
 
 import static org.assertj.core.api.Assertions.assertThat;
