+++ src/test/java/wooteco/subway/service/StationServiceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.dao.InmemoryStationDao;
+import wooteco.subway.dao.station.InmemoryStationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.NotFoundException;
